requirements.txt
fastapi==0.115.13
uvicorn==0.34.3

pydantic==2.11.7
jinja2==3.1.2
python-dotenv==1.0.0

numpy==1.26.4
pandas==2.1.1
scikit-learn==1.8.0

langchain==0.0.330
langchain-core==0.1.23
langchain-community==0.3.26
langchain-groq==0.0.1

sqlite-utils==3.36

python-multipart==0.0.9
